+++ src/graphql/event/dns.rs
@@ -3,7 +3,7 @@ use chrono::{DateTime, Utc};
 use review_database as database;
 
 use super::{country_code, find_ip_customer, find_ip_network, TriageScore};
-use crate::graphql::{customer::Customer, network::Network};
+use crate::graphql::{customer::Customer, network::Network, triage::ThreatCategory};
 
 #[allow(clippy::module_name_repetitions)]
 pub(super) struct DnsCovertChannel {
@@ -132,6 +132,10 @@ impl DnsCovertChannel {
 
     async fn confidence(&self) -> f32 {
         self.inner.confidence
+    }
+
+    async fn category(&self) -> ThreatCategory {
+        self.inner.category.into()
     }
 
     async fn triage_scores(&self) -> Option<Vec<TriageScore>> {
@@ -277,6 +281,10 @@ impl LockyRansomware {
         self.inner.confidence
     }
 
+    async fn category(&self) -> ThreatCategory {
+        self.inner.category.into()
+    }
+
     async fn triage_scores(&self) -> Option<Vec<TriageScore>> {
         self.inner
             .triage_scores
@@ -413,6 +421,10 @@ impl CryptocurrencyMiningPool {
 
     async fn coins(&self) -> Vec<String> {
         self.inner.coins.clone()
+    }
+
+    async fn category(&self) -> ThreatCategory {
+        self.inner.category.into()
     }
 
     async fn triage_scores(&self) -> Option<Vec<TriageScore>> {
@@ -553,6 +565,9 @@ impl BlockListDns {
         self.inner.ttl.clone()
     }
 
+    async fn category(&self) -> ThreatCategory {
+        self.inner.category.into()
+    }
     async fn triage_scores(&self) -> Option<Vec<TriageScore>> {
         self.inner
             .triage_scores