+++ src/graphql/event/mqtt.rs
@@ -3,7 +3,7 @@ use chrono::{DateTime, Utc};
 use review_database as database;
 
 use super::{country_code, find_ip_customer, find_ip_network, TriageScore};
-use crate::graphql::{customer::Customer, network::Network};
+use crate::graphql::{customer::Customer, network::Network, triage::ThreatCategory};
 
 pub(super) struct BlockListMqtt {
     inner: database::BlockListMqtt,
@@ -103,6 +103,10 @@ impl BlockListMqtt {
 
     async fn suback_reason(&self) -> Vec<u8> {
         self.inner.suback_reason.clone()
+    }
+
+    async fn category(&self) -> ThreatCategory {
+        self.inner.category.into()
     }
 
     async fn triage_scores(&self) -> Option<Vec<TriageScore>> {