+++ src/graphql/event/tls.rs
@@ -3,7 +3,7 @@ use chrono::{DateTime, Utc};
 use review_database as database;
 
 use super::{country_code, find_ip_customer, find_ip_network, TriageScore};
-use crate::graphql::{customer::Customer, network::Network};
+use crate::graphql::{customer::Customer, network::Network, triage::ThreatCategory};
 
 pub(super) struct BlockListTls {
     inner: database::BlockListTls,
@@ -163,6 +163,10 @@ impl BlockListTls {
 
     async fn last_alert(&self) -> u8 {
         self.inner.last_alert
+    }
+
+    async fn category(&self) -> ThreatCategory {
+        self.inner.category.into()
     }
 
     async fn triage_scores(&self) -> Option<Vec<TriageScore>> {